REFACTOR: extract navigation logic from ProductPage into useConditionalNavigation hook. FEAT: add universal hook for conditional navigation.
