+++ src/components/pages/ProductPage/ProductPage.tsx
@@ -1,7 +1,4 @@
-import { useEffect } from 'react';
-
-import { useLocation, useNavigate } from 'react-router-dom';
-
+import useConditionalNavigation from '../../../hooks/useConditionalNavigation';
 import useScrollToTop from '../../../hooks/useScrollToTop';
 
 import ProductPageContent from '../../organisms/pages/productPage/ProductPageContent/ProductPageContent';
@@ -9,21 +6,14 @@ import ProductPageWithHashData from '../../organisms/pages/productPage/ProductPa
 
 
 const ProductPage: React.FC = () => {
+
     useScrollToTop(0, 'instant');
 
-    const navigate = useNavigate()
-
-    const locationState = useLocation();
-    const data = locationState.state?.data;
-    const hash = locationState.hash;
-
-    useEffect(()=>{
-        if(!data && !hash) return navigate('/shop')
-    },[data])
+    const {data,hash} = useConditionalNavigation('/shop');
 
     if(data) return <ProductPageContent data={data}/>;
     else if(!data && hash) return <ProductPageWithHashData hash={hash.slice(1)}/>;
     else null
 }
 
-export default ProductPage;//!BLOG CARD   LATEST BLOGS?   ORDER DETAIL
+export default ProductPage;
